Fix ICE-7298 - Visual Studio 2013 build error

+++ cpp/src/Slice/Parser.cpp
@@ -5368,7 +5368,7 @@ Slice::Operation::returnsData() const
 bool
 Slice::Operation::returnsMultipleValues() const
 {
-    int count = outParameters().size();
+    size_t count = outParameters().size();
 
     if(returnType())
     {
